+++ src/sql/requirePasswordValidationPolicy.ts
@@ -1,17 +1,17 @@
 import { ResourceValidationArgs, ReportViolation, EnforcementLevel } from '@pulumi/policy'
 
-export const requireAutomatedBackup = {
-	name: 'sql-require-automated-backup',
-	description: 'Check that CloudSQL Database Instance enables automated backups.',
+export const requirePasswordValidationPolicy = {
+	name: 'sql-require-password-validation-policy',
+	description: 'Check that CloudSQL Database Instance enables password validation.',
 	enforcementLevel: 'advisory' as EnforcementLevel,
 	validateResource: (args: ResourceValidationArgs, reportViolation: ReportViolation) => {
 		if (args.type === 'gcp:sql/databaseInstance:DatabaseInstance') {
 			const settings = args.props.settings
 			if (settings) {
-				const backup = settings.backupConfiguration?.enabled
-				if (!backup) {
+				const passwordValidation = settings.passwordValidationPolicy?.enablePasswordPolicy
+				if (!passwordValidation) {
 					reportViolation(
-						'CloudSQL Database Instance should have backup configuration enabled.'
+						'CloudSQL Database Instance should have password validation policy enabled.'
 					)
 				}
 			}